backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-extensions==4.1
Werkzeug==3.1.5
django-cors-headers==4.9.0

django-stubs==5.2.9
ipdb==0.13.13
ipython==9.9.0
black==26.1.0
pydantic==2.12.5

gunicorn==24.1.1


